+++ yyskmultilearn/ext/__init__.py
@@ -1,0 +1,28 @@
+"""
+The :mod:`yyskmultilearn.ext` provides wrappers for other multi-label
+classification libraries. Currently it provides a wrapper for:
+
+Currently the available classes include:
+
++--------------------------------------------+------------------------------------------------------------------+
+| Name                                       | Description                                                      |
++============================================+==================================================================+
+| :class:`~yyskmultilearn.ext.Meka`            | Wrapper for the Multilabel Extension to WEKA -                   |
+|                                            | `MEKA <http://meka.sf.net>`_ library                             |
++--------------------------------------------+------------------------------------------------------------------+
+| :class:`~yyskmultilearn.ext.Keras`           | Wrapper for the Python Deep Learning library -                   |
+|                                            | `KERAS <http://https://keras.io/>`_                              |
++--------------------------------------------+------------------------------------------------------------------+
+| :func:`~yyskmultilearn.ext.download_meka`    | Helper function for installing MEKA                              |
++--------------------------------------------+------------------------------------------------------------------+
+
+"""
+
+import sys, platform
+from .meka import Meka, download_meka
+
+__all__ = ["Meka", 'download_meka']
+
+if not (sys.version_info[0] == 2 or platform.architecture()[0]=='32bit'):
+    from .keras import Keras
+    __all__ += ['Keras']